refactor : grouping by file type
add a components folder and moved related files under it
