+++ nearbnb/src/components/About.js
