+++ nearbnb/src/components/App.js
@@ -1,4 +1,4 @@
-import logo from './logo.svg';
+// import logo from './logo.svg';
 import './App.css';
 import Title from './Title';
 import POPOSList from './POPOSList';
