+++ nearbnb/src/components/Footer.css
