+++ nearbnb/src/components/Footer.js
