+++ nearbnb/src/components/POPOSList.css
