+++ nearbnb/src/components/POPOSList.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import POPOSSpace from './POPOSSpace';
 import './POPOSList.css';
-import data from './sfpopos-data.json';
+import data from '../sfpopos-data.json';
 
 function POPOSList() {
