+++ nearbnb/src/components/POPOSSpace.css
